backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
moviepy==1.0.3
openai-whisper==20231117
google-generativeai==0.3.2
python-dotenv==1.0.0
langchain==0.1.0
langchain-google-genai==0.0.8
pydantic==2.5.0
protobuf==4.25.1